move files & clean up
