+++ coinpurse.client/coinpurse.client.esproj
@@ -11,7 +11,4 @@
   <ItemGroup>
     <None Include=".vscode\launch.json" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="src\containers\" />
-  </ItemGroup>
 </Project>
