+++ coinpurse.client/src/Theme/Theme.tsx
@@ -1,20 +1,20 @@
-import { createTheme } from '@mui/material/styles';
-import { red } from '@mui/material/colors';
+import { createTheme } from "@mui/material/styles";
+import { red } from "@mui/material/colors";
 
 // A custom theme for this app
 const theme = createTheme({
-  cssVariables: true,
-  palette: {
-    primary: {
-      main: '#556cd6',
+    cssVariables: true,
+    palette: {
+        primary: {
+            main: "#556cd6",
+        },
+        secondary: {
+            main: "#19857b",
+        },
+        error: {
+            main: red.A400,
+        },
     },
-    secondary: {
-      main: '#19857b',
-    },
-    error: {
-      main: red.A400,
-    },
-  },
 });
 
 export default theme;
