+++ coinpurse.client/src/containers/App.tsx
