+++ coinpurse.client/src/containers/DashboardPage.tsx
