+++ coinpurse.client/src/containers/SettingsPage.tsx
