+++ coinpurse.client/src/services/accountService.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
 export const getAccounts = async () => {
-  const response = await axios.get('/api/account');
-  return response.data;
+    const response = await axios.get('/api/account');
+    return response.data;
 };
